backend/requirements.txt
Django==5.0.4
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djoser==2.2.2
Pillow==10.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7 